랭크 메세지 빌드하는 곳 옮김 - enum values() 활용해서 받아옴
